RegexState::Begin
Added the Begin method to the RegexState class.  This method retrieves the location of the regular expression engine at the time the RegexState was created.  This allows a partial rewind -- i.e. the location of the regular expression engine can be rewound by setting its current location, without any of the extra cleanup which may occur as a result of calling RegexState::Rewind.