Engine Backtracking Fix
Fixed a backtracking error which was causing an infinite loop under certain circumstances.

When a RegexEngine that has previously been invoked is invoked again (to try and match a different substring) it works backwards through the RegexState objects associated with its RegexPatternElements.  This involves moving an iterator into the collection of RegexPatternElements backwards as each RegexState is considered and discarded.

However, due to an off by one error, the iterator into the collection of RegexPatternElements would always point to the element after the intended element, which meant that when retrying the last RegexPatternElement that was matched, begin==end, and an infinite loop would occur as the RegexEngine would skip the matching process and return the same match it had formed previously.