Multiple Fix
Fixed a bug which was causing an infinite loop under certain circumstances.

When a RegexMultiple is invoked during backtracking, it backtracks internally, walking over all the matches it's made of its internal RegexPatternElement.

In the case where the last match can backtrack -- i.e. can possibly match a different substring -- the number of RegexStates in the RegexMultiple's private state does not change (since backtracking considers the last RegexState, finds that its pattern element can be invoked again, and ends without deleting it).  In the instance where the RegexMultiple has a maximum number of matches that it can form, this was misinterpreted as the match already being done (since the number of RegexStates was equal to the maximum number of matches the RegexMultiple was permitted to form), meaning that the RegexMultiple would repeatedly match the same substring.

Added a check so the comparison between the maximum number of matches (if applicable) and the number of RegexState objects in the RegexMultiple's private state would not occur immediately after backtracking, rectifying this issue.