Regular Expression Unit Tests
Started adding unit tests for the regular expression engine.

Split the String and CodePoint Catch::toString overloads into their own header and source file so that they may be shared without copying and pasting.

The regular expression unit tests will be split across multiple files all under src/test/regex to reduce compilation times.  It is expected that there will be more scenarios/asserts for the regular expression engine than there currently are for the rest of the library, and the unit tests for the rest of the library already take a long time to compile such that it is tedious recompiling them iteratively as one works on them.

Broke the Catch boilerplate code which generates the main method out into src/test/main.cpp, renaming the previous src/test/main.cpp (which contains all the unit tests for the non-regex parts of the library) to src/test/main_old.cpp.  At some point the unit tests for the rest of the library will be refactored to follow the multiple file model being used by the regular expression unit tests, and src/test/main_old.cpp will be removed.

Currently have unit tests for ^, $, \A, \Z, \z, \G, \b, and \B -- i.e. the anchoring constructs.