Lazy Repetition
Implemented support for lazy repetition (e.g. a*?).  Previously there was support for parsing lazy repetition structures, but there was no implementation.

Lazy repetition is realized as progressive greedy repetition.  Accordingly the greedy repetition driver has been refactored into the RegexMultipleState class and is now shared.  RegexLazyMultiple uses this driver, but limits the minimum and maximum number of matches which may be formed, starting at the minimum and working upwards to the maximum as matches fail or as the engine backtracks and requires that the RegexLazyMultiple form different matches.