Split
Added the Regex::Split and Regex::SplitIterate methods and supporting infrastructure, which allow a string to be split into substrings which are bounded by the beginning and end of the string, and each match of the regular expression.

SplitIterate leverages the Regex::SplitIterable and Regex::SplitIterator classes, which allow the string to be lazily split, much like Iterate/Regex::Iterable/Regex::Iterator do for matching.