Character Classes Unit Tests
Added unit tests for the \w, \W, \s, \S, \d, and \D constructs.

All unit test scenarios in characterclasses.cpp have the "[regex]" tag -- to be added to all other regular expression unit tests scenarios in the future -- in anticipation of a time when all unit tests for the library are merged into one executable.