Group Recursion
Added support for group recursion constructs from PCRE.  These take the form of:

- (?n), where n is an integer
- (?&name)
- (?P>name)

and cause the engine to execute the referenced group again.

The (?R) recursion construct -- which causes recursion of the entire pattern -- is not yet supported as the RegexEngine class will have to be modified to support this (as the pattern element will need to be able to acquire a reference to this entire pattern).

The (?+n) and (?-n) recursion constructs -- which recurse a group based on its relative location -- are not yet supported as the RegexCompiler class will have to be modified to support this.