RegexEngine Entire Pattern Reference
RegexEngine objects now contain a member -- Pattern -- which always refers to the outermost pattern.  When a RegexEngine is initialized without using an existing RegexEngine this member is set to refer to the pattern that RegexEngine matches.  When a RegexEngine is initialized using an existing RegexEngine and a new pattern (as when matching a subpattern) this member continues to refer to the outermost pattern.  In this way pattern elements always have access to the entire, outermost pattern.