Recurse Entire Pattern
Added support for the (?R) construct, which recurses the entire pattern.

The implementation and parsing for this construct are different enough that I felt they didn't belong in the same file, so group/subpattern recursion has been moved to regexrecursegroup.cpp, whereas recursion of the entire pattern is in regexrecurse.cpp.