Backreferences
Added support and unit tests for the backreferencing constructs:

- \number
- \k<name>
- \k<number>

The code for the backreference pattern elements and parser is in regexliteral.cpp so that it may share the literal matching code written for RegexLiteral.  This code has been refactored into a base class -- RegexLiteralBase -- which is shared between RegexLiteral and RegexBackreference.