RegexMatch Numeric Capturing Group Names
When the RegexMatch class is supplied with a String to its operator [] it will now check if the given string is numeric, and, if it is, will look up the corresponding number instead.

This is the same transform as is performed during compilation when a capturing group is explicitly named -- numeric strings are transformed into numbers.