Group Overhaul and Unit Tests
Added unit tests for non-capturing groups, capturing groups (named, numbered, and automatically numbered), and non-capturing groups which change the engine's options.

RegexBalancing now properly derives from RegexGroup, which simplifies its implementation and reduces repeated code.

Substantially rewrote the parsing/compiling code for grouping constructs for simplicity, and in so doing fixed a small bug:

When compiling a capturing group, the compiler would compile the entire captured subpattern before adding the pattern element to the RegexCompiler's index of capturing groups.  This meant that if a capturing group with an identical name or number was nested, they would be added out of order.