Add Capturing Group Unit Tests
Added unit tests for multiple captures -- i.e. when the same group captures multiple times in a single match.  Verifies that captures are made left-to-right in left-to-right mode, and right-to-left in right-to-left mode.

Changed another scenario so that assertions wouldn't be double counted -- nesting THEN blocks was causing Catch to double count assertions in the outer block.